feat: refactor mock handlers by separating Twitch and internal handlers
