+++ src/mocks/handlers/internal.ts
@@ -1,0 +1,3 @@
+import { http, HttpResponse } from 'msw';
+
+export const handlers = [];
